comments can be posted. Delete animal added

+++ JavaProject/src/main/java/com/ENSF607/AnimalProject/repository/CommentRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment,Integer> {
+public interface CommentRepository extends JpaRepository<Comment,Long> {
     List<Comment> findAll();
 
-    List<Comment> findAllByanimalid(Integer animalid);
+    List<Comment> findAllByanimalid(Long animalid);
 
 }
